backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
openpyxl==3.1.2
pandas==2.2.2
